Various
Added function mycaxis, to be called instead of caxis. caxis is deprecated from Matlab 2022b onwards, in favour of clim, which is more strict about argument-checking. mycaxis is designed to be compatible with all versions from 2021b onwards. It sanitises its crange argument, and calls clim if present, and caxis if not.

GFtbox models have a new plot option, 'edgesharpness'. For volumetric meshes, this determines which surface edges should be drawn when the 'drawedges' plot option is 1, by drawing all surface edges at which the surface has an angle above the threshold defined by 'edgesharpness'. By default no edges are drawn.